+++ class/Themes/Wonder/Form/Components/File.php
@@ -21,7 +21,7 @@ class File extends Field
         $script = $this->renderDataTransferScript($id);
 
         return <<<HTML
-<div class="{$this->containerClass('file')}">
+<div class="{$this->containerClass('file')} compiled">
     <label for="{$id}" class="wi-label">{$label}</label>
     <input class="wi-input" id="{$id}" type="file" accept="{$accept}" name="{$nameArray}" data-wi-max-file="{$maxFile}" data-wi-max-size="{$maxSize}" data-wi-check="true"{$this->labelMarker()} {$multiple} {$attributes}>
     {$this->renderError()}